requirements.txt
aws-cdk-lib==2.83.1
constructs>=10.0.0,<11.0.0
aws-cdk.aws-neptune-alpha==2.83.1a0
